Add pagination and infinite scrolling

+++ src/commonMain/kotlin/jezorko/github/gfngameslist/games/GetGamesResponse.kt
@@ -4,7 +4,7 @@ import kotlinx.serialization.Serializable
 
 @Serializable
 data class GetGamesResponse(
-    val supportedGamesCount: Long,
+    val supportedGamesCount: Int,
     val lastUpdatedAt: Long,
     val games: List<Game>
 )
